+++ packages/@expo/cli/CHANGELOG.md
@@ -13,6 +13,7 @@
 ### 💡 Others
 
 - Allowing import to ESM modules. ([#39390](https://github.com/expo/expo/pull/39390) by [@kudo](https://github.com/kudo))
+- Added MCP tunnel integration. ([#39392](https://github.com/expo/expo/pull/39392) by [@kudo](https://github.com/kudo))
 
 ## 0.26.8 — 2025-09-03